factored out header seach bar
