+++ src/pages/profile/index.tsx
@@ -1,6 +1,6 @@
+import { HeaderSearchBar, SEARCH_QUERY_KEY } from "@/components/HeaderSearchBar";
 import { HeaderLink, Layout } from "@/components/Layout";
 import { Clickable, Star, UserPic } from "@/components/UI";
-import { hasFocusWithin } from "@/hooks";
 import { getQueryParam, includes, minutesToMs, setQueryParam } from "@/utils";
 import { api, type RouterInputs, type RouterOutputs } from "@/utils/api";
 import * as Dialog from "@radix-ui/react-dialog";
@@ -11,22 +11,32 @@ import Image from "next/image";
 import Link from "next/link";
 import type { NextRouter } from "next/router";
 import { useRouter } from "next/router";
-import { useEffect, useRef, useState, type RefObject } from "react";
-import { createPortal } from "react-dom";
+import { useEffect, useRef, type RefObject } from "react";
 import { Flipped, Flipper } from "react-flip-toolkit";
 import EggBasketIcon from "~icons/custom/egg-basket.jsx";
 import GroceriesIcon from "~icons/custom/groceries.jsx";
 import MilkIcon from "~icons/custom/milk.jsx";
-import SearchIcon from "~icons/iconamoon/search.jsx";
 import SwapIcon from "~icons/iconamoon/swap.jsx";
-import ResetIcon from "~icons/radix-icons/cross-1";
 import SettingsIcon from "~icons/solar/settings-linear";
 
 export default function Page() {
   const { data, status } = useSession();
 
   return (
-    <Layout header={{ header: <HeaderFilterInput /> }}>
+    <Layout
+      header={{
+        header: (
+          <HeaderSearchBar
+            right={
+              <HeaderLink
+                Icon={SettingsIcon}
+                href="/profile/settings"
+              />
+            }
+          />
+        ),
+      }}
+    >
       {status === "authenticated" ? (
         <div className="flex w-full flex-col gap-2 p-4">
           <ProfileInfo user={data.user} />
@@ -69,89 +79,6 @@ function Reviews() {
       </div>
       {/* That way we fetch ReviewCards w/o waiting for countQuery to settle */}
       {!countQuery.isSuccess || !!countQuery.data ? <ReviewCards /> : <NoReviews />}
-    </div>
-  );
-}
-
-const filterKey = "search";
-function HeaderFilterInput() {
-  const [isOpen, setIsOpen] = useState(false);
-  const debounceTimeoutRef = useRef<number>();
-
-  const router = useRouter();
-  const filterParam: string = getQueryParam(router.query[filterKey]) ?? "";
-  const [filter, setFilter] = useState(filterParam);
-
-  const searchIcon = <SearchIcon className="h-7 w-7" />;
-
-  // keeps filter in sync on back/forward
-  useEffect(() => {
-    setFilter(filterParam);
-  }, [filterParam, setFilter]);
-
-  return (
-    <div
-      className="flex h-full items-center gap-2 text-xl"
-      onBlur={hasFocusWithin((hasFocus) => {
-        if (!isOpen) return;
-        setIsOpen(hasFocus);
-      })}
-    >
-      {isOpen ? (
-        <>
-          {createPortal(
-            <div className="absolute inset-0 z-0 animate-fade-in bg-black/50" />,
-            document.body
-          )}
-          {searchIcon}
-          <input
-            // key helps refocus input when clear button is pressed
-            key={`${!!filter}`}
-            autoFocus
-            className="h-full grow p-1 caret-app-green outline-none"
-            placeholder="Seach by..."
-            value={filter}
-            onChange={(e) => {
-              const { value } = e.target;
-              setFilter(value);
-
-              window.clearTimeout(debounceTimeoutRef.current);
-              debounceTimeoutRef.current = window.setTimeout(() => {
-                setQueryParam(router, filterKey, value);
-              }, 1000);
-            }}
-          />
-          <button
-            aria-label="reset filter"
-            className="ml-1"
-            onClick={() => {
-              setFilter("");
-              setIsOpen(false);
-
-              window.clearTimeout(debounceTimeoutRef.current);
-              setQueryParam(router, filterKey, null);
-            }}
-          >
-            <ResetIcon className="h-7 w-7" />
-          </button>
-        </>
-      ) : (
-        <>
-          <button
-            aria-label="Start review search"
-            onClick={() => {
-              setIsOpen(true);
-            }}
-          >
-            {searchIcon}
-          </button>
-          <div className="grow text-center">Profile</div>
-          <HeaderLink
-            Icon={SettingsIcon}
-            href="/profile/settings"
-          />
-        </>
-      )}
     </div>
   );
 }
@@ -238,7 +165,7 @@ function parseSortParam(param: SortOption): SortQuery {
 const limit = 20;
 function ReviewCards() {
   const router = useRouter();
-  const filter = getQueryParam(router.query[filterKey]);
+  const filter = getQueryParam(router.query[SEARCH_QUERY_KEY]);
   const sortParam = useParseSort(router);
   const sort = parseSortParam(sortParam);
