+++ src/pages/search.tsx
@@ -1,8 +1,23 @@
-import { Layout } from "@/components/Layout";
+import { HeaderSearchBar } from "@/components/HeaderSearchBar";
+import { HeaderLink, Layout } from "@/components/Layout";
+import HomeIcon from "~icons/uil/home-alt";
 
 export default function Page() {
   return (
-    <Layout header={{ title: "Search" }}>
+    <Layout
+      header={{
+        header: (
+          <HeaderSearchBar
+            right={
+              <HeaderLink
+                Icon={HomeIcon}
+                href="/"
+              />
+            }
+          />
+        ),
+      }}
+    >
       <div>SEARCH PAGE</div>
     </Layout>
   );
